fix: semantics error on how the timezone is translated and how the timestamp is stored pn database

+++ app/src/main/java/com/gntr/professionalsleeper/data/local/dao/SleepSessionDao.kt
@@ -23,7 +23,7 @@ interface SleepSessionDao {
     suspend fun getAllSessionsSnapshot(): List<SleepSession>
 
     @Query("SELECT * FROM sleep_sessions WHERE id = :id")
-    suspend fun getSessionById(id: Int): SleepSession?
+    suspend fun getSessionById(id: Long): SleepSession?
 
     @Query("DELETE FROM sleep_sessions")
     suspend fun deleteAllSessions()
